requirements.txt
aiohttp
async-timeout
azure-identity
beautifulsoup4
fastapi
jinja2
openai
python-multipart
uvicorn
sqlalchemy>=2.0
psycopg2-binary>=2.9
schedule
pika==1.3.1
PyGithub
markdown
alembic
prometheus-client>=0.19.0
redis>=4.0.0
